refactor: 채팅방 헤더 수정 #136
