+++ apps/web/src/components/layout/index.ts
@@ -3,7 +3,6 @@ export { default as MainHeader } from './header/main';
 export { default as AuctionHeader } from './header/auction';
 export { default as SearchHeader } from './header/search';
 export { default as ChatHeader } from './header/chat';
-export { default as ChatRoomHeader } from './header/chat-room';
 export { default as ProfileHeader } from './header/profile';
 
 export { default as Navigation } from './navigation';
